Thorough re-organization
