--- EngineCode Viewer/Engine_code/Shaders/Colour Map/Fragment.txt
@@ -1,8 +1,0 @@
-#version 400 core
-in vec3 x_Color;
-out vec4 FragColor;
-
-
-void main() {
-FragColor = vec4(x_Color, 1.0);
-}
