--- EngineCode Viewer/Engine_code/Shaders/Colour Map/Vertex.txt
@@ -1,20 +1,0 @@
-# version 330 core
-layout(location = 0) in vec3 a_position;
-layout(location = 1) in vec3 a_Color;
-
-
-out vec3 x_Color;
-
-uniform mat4 projection;
-uniform mat4 view;
-uniform int scale;
-
-void main()
-{
-    x_Color = a_Color;
-    gl_Position = projection * view * vec4(a_position, 1.0);
-
-    vec4 cs_position = view * vec4(a_position, 1.0);
-    gl_PointSize =  -1/(cs_position.z)*scale;
-
-}
