--- EngineCode Viewer/Engine_code/Shaders/Height Map/Fragment.txt
@@ -1,10 +1,0 @@
-#version 400
-
-in float height;
-uniform float max;
-
-void main() {
- float normalizedHeight = height/max;
- vec4 color = vec4(normalizedHeight,normalizedHeight, normalizedHeight,1.0);
- gl_FragColor = color;
-}
