--- EngineCode Viewer/Engine_code/Shaders/Height Map/Vertex.txt
@@ -1,18 +1,0 @@
-# version 330 core
-layout(location = 0) in vec3 a_position;
-
-uniform mat4 projection;
-uniform mat4 view;
-uniform int scale;
-
-out float height;
-
-void main()
-{
-    gl_Position = projection * view * vec4(a_position, 1.0);
-
-    vec4 cs_position = view * vec4(a_position, 1.0);
-    gl_PointSize =  -1/(cs_position.z)*scale;
-
-    height = a_position.z;
-}
