+++ Open3d Viewer/bound.json
@@ -1,0 +1,18 @@
+
+    {
+    "axis_max": 150000,
+    "axis_min": 25000,
+    "bounding_polygon": [
+        [-999750.0, 999750.0, 0.0],
+        [999750.0, 999750.0, 0.0],
+        [999750.0, -999750.0, 0.0],
+        [-999750.0, -999750.0, 0.0]
+    ],
+    "class_name": "SelectionPolygonVolume",
+    "orthogonal_axis": "Z",
+    "version_major": 1,
+    "version_minor": 0
+    }
+
+
+
