--- bound.json
@@ -1,18 +1,0 @@
-
-    {
-    "axis_max": 10000000,
-    "axis_min": -10000000,
-    "bounding_polygon": [
-        [-500000.0, 500000.0, 0.0],
-        [500000.0, 500000.0, 0.0],
-        [500000.0, -500000.0, 0.0],
-        [-500000.0, -500000.0, 0.0]
-    ],
-    "class_name": "SelectionPolygonVolume",
-    "orthogonal_axis": "Z",
-    "version_major": 1,
-    "version_minor": 0
-    }
-
-
-
